+++ aimmo-game/simulation/map_generator.py
@@ -7,7 +7,7 @@ from itertools import tee
 from six.moves import zip, range
 
 from simulation.direction import ALL_DIRECTIONS
-from simulation.location import Location
+from simulation.geography.location import Location
 from simulation.world_map import WorldMap
 
 from simulation.custom_map import BaseGenerator
