+++ aimmo-game/tests/test_service.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from unittest import TestCase, skip
 
 from simulation.game_state import GameState
-from simulation.location import Location
+from simulation.geography.location import Location
 from simulation.turn_manager import state_provider
 from simulation.world_map import WorldMap
 from simulation.world_state import WorldState
