+++ aimmo-game/tests/test_simulation/avatar/test_avatar_view.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from simulation.geography.cell import Cell
-from simulation.location import Location
+from simulation.geography.location import Location
 from simulation.direction import EAST, WEST, NORTH, SOUTH
 from simulation.avatar.avatar_view import AvatarView
