+++ aimmo-game/tests/test_simulation/maps.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 
 from collections import defaultdict
 
-from simulation.location import Location
+from simulation.geography.location import Location
 from simulation.world_map import WorldMap
 from simulation.geography.cell import Cell
