+++ aimmo-game/tests/test_simulation/test_action.py
@@ -7,7 +7,7 @@ from simulation import event
 from simulation.avatar.avatar_manager import AvatarManager
 from simulation.direction import EAST
 from simulation.game_state import GameState
-from simulation.location import Location
+from simulation.geography.location import Location
 from .dummy_avatar import MoveDummy
 from .maps import InfiniteMap, EmptyMap, AvatarMap
