+++ aimmo-game/tests/test_simulation/test_custom_map.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 import random
 import unittest
 
-from simulation.location import Location
+from simulation.geography.location import Location
 from simulation.pickups import DeliveryPickup
 from simulation.world_map import WorldMap
 from simulation.custom_map import ScoreCellDecoder
