+++ aimmo-game/tests/test_simulation/test_game_state.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from unittest import TestCase
 
 from simulation.game_state import GameState
-from simulation.location import Location
+from simulation.geography.location import Location
 from .dummy_avatar import DummyAvatar
 from .dummy_avatar import DummyAvatarManager
 from .maps import InfiniteMap, AvatarMap, EmptyMap
