+++ aimmo-game/tests/test_simulation/test_map_generator.py
@@ -4,7 +4,7 @@ import random
 import unittest
 
 from simulation import map_generator
-from simulation.location import Location
+from simulation.geography.location import Location
 from simulation.map_generator import get_random_edge_index
 from simulation.world_map import WorldMap
 from .dummy_avatar import DummyAvatarManager
