+++ aimmo-game/tests/test_simulation/test_world_map.py
@@ -5,7 +5,7 @@ from string import ascii_uppercase
 
 from unittest import TestCase
 
-from simulation.location import Location
+from simulation.geography.location import Location
 from simulation.world_map import WorldMap, world_map_static_spawn_decorator
 from simulation.geography.cell import Cell
 from .dummy_avatar import DummyAvatar
